project_selenium/requirements.txt
pytest
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure
allure-pytest
selenium
webdriver-manager
